video branch: Fixed audio timestamps in AudioSource.
svn path=/branches/libavg_video/; revision=8586

+++ src/audio/AudioSource.h
@@ -49,7 +49,7 @@ private:
     AudioMsgQueue& m_StatusQ;
     int m_SampleRate;
     AudioBufferPtr m_pInputAudioBuffer;
-    int m_LastTime;
+    float m_LastTime;
     int m_CurInputAudioPos;
     bool m_bPaused;
 };
